refactor: move SF layout utilities from modified_utils/ into blackwell_helpers.py
BlockScaledBasicChunk, make_smem_layout_sfa/sfb, make_tmem_layout_sfa/sfb
now live alongside other block-scaled MMA helpers. Removes the
modified_utils/ directory entirely.
